feat(site) See GitHub API Rate limit (#2308)
* See API Rate limit

* Skip tests

* See rate

* Restore fetch script

* log token

* JSONify res

* clear console

* Refactor script fetching

* Revert back test script on deploy script

* Use lowercase name in filename check

* Rename file
