+++ src/scripts/fetch_packages.js
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
-// ./fetch_package_files <file> <output> < input
-// ./fetch_package_files "README.md" "./output" < input
+// ./fetch_package_names <suffix> > output
+// ./fetch_package_names "-loader" > output.json
+const GitHubApi = require('github');
 const fs = require('fs');
 const urlModule = require('url');
 const path = require('path');
@@ -10,52 +11,99 @@ const request = require('request');
 const _ = require('lodash');
 
 if (require.main === module) {
-    main();
+  main();
 } else {
-    module.exports = fetchPackageFiles;
+  module.exports = fetchPackageNames;
 }
 
 function main() {
-  const file = process.argv[2];
-  const output = process.argv[3];
+  const organization = process.argv[2];
+  const suffix = process.argv[3];
+  const file = process.argv[4];
+  const output = process.argv[5];
 
-  if ( !file ) {
+  if (!organization) {
+    return console.error('Missing organization!');
+  }
+  if (!suffix) {
+    return console.error('Missing suffix!');
+  }
+
+  if (!file) {
     return console.error('Missing file!');
   }
 
-  if ( !output ) {
+  if (!output) {
     return console.error('Missing output!');
   }
 
   mkdirp.sync(output);
 
-  const stdin = process.openStdin();
-  var input = '';
-
-  stdin.setEncoding('utf8');
-  stdin.on('data', function(d) {
-    input += d;
-  });
-
-  stdin.on('end', function() {
-    fetchPackageFiles({
-      input: JSON.parse(input),
-      file: file,
-      output: path.resolve(process.cwd(), output),
-      limit: 4
-    }, function(err, d) {
+  fetchPackageNames(
+    {
+      organization: organization,
+      suffix: suffix
+    },
+    function(err, d) {
       if (err) {
         return console.error(err);
       }
 
-      const msg = d.length === 0
-        ? 'Fetched 0 files'
-        : d.length === 1
-        ? 'Fetched 1 file: '
-        : `Fetched ${d.length} files: `;
-      console.log(msg + _.map(d, 'full_name'));
+      fetchPackageFiles(
+        {
+          input: d,
+          file: file,
+          output: path.resolve(process.cwd(), output),
+          limit: 4
+        },
+        function(err, d) {
+          if (err) {
+            return console.error(err);
+          }
+
+          const msg =
+            d.length === 0 ? 'Fetched 0 files' : d.length === 1 ? 'Fetched 1 file: ' : `Fetched ${d.length} files: `;
+          console.log(msg + _.map(d, 'full_name'));
+        }
+      );
+    }
+  );
+}
+
+function fetchPackageNames(options, cb) {
+  const github = new GitHubApi();
+
+  if (process.env.GITHUB_TOKEN) {
+    github.authenticate({
+      type: 'token',
+      token: process.env.GITHUB_TOKEN
     });
-  });
+
+    github.misc.getRateLimit({}, (err, res) => {
+      if (err) throw err;
+      console.log(res.data.rate);
+    });
+  }
+
+  // XXX: weak since this handles only one page
+  github.repos.getForOrg(
+    {
+      org: options.organization,
+      per_page: 100
+    },
+    function(err, d) {
+      if (err) {
+        return cb(err);
+      }
+
+      return cb(
+        null,
+        d.data.filter(function(o) {
+          return o.name.endsWith(options.suffix);
+        })
+      );
+    }
+  );
 }
 
 function fetchPackageFiles(options, finalCb) {
@@ -73,7 +121,7 @@ function fetchPackageFiles(options, finalCb) {
           return cb(err);
         }
 
-        if (body && file === 'README.md') {
+        if (body && file.toLowerCase() === 'readme.md') {
           body = body
             // Remove all but the description from lead-in (XXX: Optional `\/?` should be unnecessary)
             .replace(/[^]*?<div align="center">([^]*?)<\/div>/, (match, content) => {
@@ -88,10 +136,19 @@ function fetchPackageFiles(options, finalCb) {
             // - [click here](LICENSE) --> [click here](https://raw.githubusercontent.com/user/repository/branch/LICENSE)
             // - [click here](./LICENSE) --> [click here](https://raw.githubusercontent.com/user/repository/branch/LICENSE)
             // - [click here](#LICENSE) --> [click here](#LICENSE)
-            .replace(/\[([^[\]]*)\]\(([^)#]+)\)/g, (match, textContent, href) => `[${textContent}](${urlModule.resolve(url, href)})`)
+            .replace(
+              /\[([^[\]]*)\]\(([^)#]+)\)/g,
+              (match, textContent, href) => `[${textContent}](${urlModule.resolve(url, href)})`
+            )
             // Modify links to keep them within the site
-            .replace(/https?:\/\/github.com\/(webpack|webpack-contrib)\/([-A-za-z0-9]+-loader\/?)([)"])/g, '/loaders/$2/$3')
-            .replace(/https?:\/\/github.com\/(webpack|webpack-contrib)\/([-A-za-z0-9]+-plugin\/?)([)"])/g, '/plugins/$2/$3')
+            .replace(
+              /https?:\/\/github.com\/(webpack|webpack-contrib)\/([-A-za-z0-9]+-loader\/?)([)"])/g,
+              '/loaders/$2/$3'
+            )
+            .replace(
+              /https?:\/\/github.com\/(webpack|webpack-contrib)\/([-A-za-z0-9]+-plugin\/?)([)"])/g,
+              '/plugins/$2/$3'
+            )
             // Replace any <h2> with `##`
             .replace(/<h2[^>]*>/g, '## ')
             .replace(/<\/h2>/g, '')
@@ -116,14 +173,8 @@ function fetchPackageFiles(options, finalCb) {
 
         return async.parallel(
           [
-            fs.writeFile.bind(null,
-              path.resolve(options.output, pkg.name + path.extname(file)),
-              headmatter + body
-            ),
-            fs.writeFile.bind(null,
-              path.resolve(options.output, pkg.name + '.json'),
-              JSON.stringify(pkg, null, 2)
-            )
+            fs.writeFile.bind(null, path.resolve(options.output, pkg.name + path.extname(file)), headmatter + body),
+            fs.writeFile.bind(null, path.resolve(options.output, pkg.name + '.json'), JSON.stringify(pkg, null, 2))
           ],
           function(err) {
             if (err) {
